20170222-03

+++ downloader.sh
@@ -4,12 +4,11 @@
 # transcode into another format
 # upload to new resource
 # 
-# Arguments:  unpacked_in_hex_filename  [ transcode_to_format ]
+# Arguments:  id
 # or
-# Arguments:  filename [ transcode_to_format ]
-# eg downloader.sh  abc.avi
-# also possible 
-# eg downloader.sh  012345abcdef 
+# Arguments:  id
+# eg downloader.sh  123
+
 
 BASENAME=`basename $0`
 cd `dirname $0`
@@ -19,25 +18,15 @@ source "$DIRNAME/common.sh"
 #DEBUG=1
 
 ID=$1
-HEX_FILENAME=$2
-TRANSCODE_FORMAT=$3
-
 WORKING_DIR=$DATA_DIR/$ID
 PROCESS_LOG=$WORKING_DIR/$$.log
-#TRANSCODE_FORMATS_LIST='320' 
 TRANSCODE_FORMATS_LIST='320 640 400 700 1100 1300 1500 175k' 
 
-
 # check the arguments
-if [[ "x$ID" == "x" ||  "x$HEX_FILENAME" == "x"   ]] ; then
-	echo "Usage:$0 id file.mp4  [transcode_to_format]"
+if [[ "x$ID" == "x"  ]] ; then
+	echo "Usage:$0 id "
 	exit 1
 fi	
-
-
-if [ "x$TRANSCODE_FORMAT" != "x" ] ; then
-	TRANSCODE_FORMATS_LIST=$TRANSCODE_FORMAT
-fi
 
 if [ "x$DEBUG" == "x1" ]; then
 	echo "mkdir -p '$WORKING_DIR'"
@@ -45,30 +34,38 @@ else
 	[ -d "$WORKING_DIR" ] || mkdir -p "$WORKING_DIR"
 fi	
 
+JOB_SETTINGS_FILE=${WORKING_DIR}/job_settings.sh
+if [ ! -f $JOB_SETTINGS_FILE ]; then
+	w2log "File $JOB_SETTINGS_FILE do not exist. Cannot set parameters"
+	exit 1
+fi
+source $JOB_SETTINGS_FILE
+
+
 DATE=`date +%Y-%m-%d_%H:%M:%S`
 MY_PID_FILE="${WORKING_DIR}/$$.downloader.pid"
 echo  "$$"  > $MY_PID_FILE
 
 # check if filename in simple text or "hex packed" ( check if found '.' in string)
-echo $HEX_FILENAME | grep "\." >/dev/null 2>&1
-if [  $? -ne 0  ]; then
-	FILENAME=`echo $HEX_FILENAME | /usr/bin/perl -ne 'print pack( "h*",$_);'`
-else
-	FILENAME=$HEX_FILENAME
-fi	
+#echo $HEX_FILENAME | grep "\." >/dev/null 2>&1
+#if [  $? -ne 0  ]; then
+#	FILENAME=`echo $HEX_FILENAME | /usr/bin/perl -ne 'print pack( "h*",$_);'`
+#else
+#	FILENAME=$HEX_FILENAME
+#fi	
 
-OUTPUT_FILENAME=`echo $FILENAME | /usr/bin/perl -ne '/^(.+)\.(\w+)$/; my $filename=$1; my $ext=$2; $filename=~s/\W/_/g; print "$filename.$ext";'`
+#OUTPUT_FILENAME=`echo $FILENAME | /usr/bin/perl -ne '/^(.+)\.(\w+)$/; my $filename=$1; my $ext=$2; $filename=~s/\W/_/g; print "$filename.$ext";'`
 # output filename with fixed special chars and spaces
 
 if [ "x$DEBUG" == "x1" ]; then
-	echo timeout ${TIMEOUT_GET_FILE} $RSYNC "${REMOTE_SOURCE}/${FILENAME}" ${WORKING_DIR}/${OUTPUT_FILENAME}
+	echo timeout ${TIMEOUT_GET_FILE} $RSYNC "${REMOTE_SOURCE_FILE}" ${DOWNLOAD_TO}
 else
-	w2log "Start download '${REMOTE_SOURCE}/${FILENAME}' and save to '$WORKING_DIR/${OUTPUT_FILENAME}'"
-	timeout ${TIMEOUT_GET_FILE} $RSYNC "${REMOTE_SOURCE}/${FILENAME}" ${WORKING_DIR}/${OUTPUT_FILENAME} >> $PROCESS_LOG 2>&1
+	w2log "Start download '${RREMOTE_SOURCE_FILE}' and save to '${DOWNLOAD_TO}'"
+	timeout ${TIMEOUT_GET_FILE} $RSYNC "${REMOTE_SOURCE_FILE}" ${DOWNLOAD_TO} >> $PROCESS_LOG 2>&1
 fi
 
 if [  $? -ne 0  ]; then
-	w2log "Error: Cannot get file '${REMOTE_SOURCE}/${FILENAME}'. See $PROCESS_LOG"
+	w2log "Error: Cannot get file '${REMOTE_SOURCE_FILE}'. See $PROCESS_LOG"
 	rm -rf $MY_PID_FILE
 	exit 1
 fi	
@@ -76,8 +73,8 @@ fi
 
 
 if [ "x$DEBUG" != "x1" ]; then
-	if [ ! -f "$WORKING_DIR/${OUTPUT_FILENAME}"  ]; then
-		w2log "File $WORKING_DIR/${OUTPUT_FILENAME} do not exist"
+	if [ ! -f "${DOWNLOAD_TO}"  ]; then
+		w2log "File ${DOWNLOAD_TO} do not exist"
 		rm -rf $MY_PID_FILE
 		exit 1
 	fi	
@@ -85,32 +82,31 @@ fi
 
 
 for i in $TRANSCODE_FORMATS_LIST; do
-	SOURCE_FILENAME="${OUTPUT_FILENAME}.${i}"
-	CMD="ln ${WORKING_DIR}/${OUTPUT_FILENAME} ${WORKING_DIR}/${SOURCE_FILENAME}"
+	CMD="ln ${DOWNLOAD_TO} ${DOWNLOAD_TO}.${i}"
 	if [ "x$DEBUG" == "x1" ]; then
 		echo $CMD 
 	else
-		w2log "Make hard link ${WORKING_DIR}/${OUTPUT_FILENAME} to ${WORKING_DIR}/${SOURCE_FILENAME}"
-		[ -f ${WORKING_DIR}/${SOURCE_FILENAME} ] || $CMD >> $PROCESS_LOG 2>&1
+		w2log "Make hard link ${DOWNLOAD_TO} ${DOWNLOAD_TO}.${i}"
+		[ -f "${DOWNLOAD_TO}.${i}" ] || $CMD >> $PROCESS_LOG 2>&1
 		if [  $? -ne 0  ]; then
-			w2log "Error: Faied make hard link ${WORKING_DIR}/${OUTPUT_FILENAME} to ${WORKING_DIR}/${SOURCE_FILENAME}"
+			w2log "Error: Faied make hard link ${DOWNLOAD_TO} to ${DOWNLOAD_TO}.${i}"
 			continue
 		fi			
 	fi
 	if [ "x$DEBUG" == "x1" ]; then
-		echo ${DIRNAME}/send2queue.pl transcoder "${DIRNAME}/transcoder.sh $ID ${WORKING_DIR}/${SOURCE_FILENAME}  $i"
+		echo ${DIRNAME}/send2queue.pl transcoder "${DIRNAME}/transcoder.sh $ID $i"
 	else
-		w2log "Put job in queue transcoder: ${DIRNAME}/transcoder.sh $ID ${WORKING_DIR}/${SOURCE_FILENAME}  $i"
-		${DIRNAME}/send2queue.pl transcoder "${DIRNAME}/transcoder.sh $ID ${WORKING_DIR}/${SOURCE_FILENAME}  $i" >> ${PROCESS_LOG} 2>&1
+		w2log "Put job in queue transcoder: ${DIRNAME}/transcoder.sh $ID $i"
+		${DIRNAME}/send2queue.pl transcoder "${DIRNAME}/transcoder.sh $ID $i" >> ${PROCESS_LOG} 2>&1
 	fi
 	if [  $? -ne 0  ]; then
 		w2log "Error: Cannot put job in queue. See $PROCESS_LOG"
-		rm -rf ${WORKING_DIR}/${SOURCE_FILENAME}
+		rm -rf "${DOWNLOAD_TO}.${i}"
 		continue
 	fi	
 done
 
 #w2log "Process $$ finished successfully"
-rm -rf ${WORKING_DIR}/${OUTPUT_FILENAME}
+rm -rf ${DOWNLOAD_TO}
 rm -rf $MY_PID_FILE
 exit 0
